# openapi.yaml
openapi: 3.0.0
info:
  title: Where in the World — Leaderboard API
  description: REST API for a geography trivia game leaderboard with scoring, statistics, and performance tracking.
  version: 1.0.0

servers:
  - url: https://your-railway-url.up.railway.app
    description: Production server
  - url: http://localhost:3000
    description: Local development

paths:
  /add:
    post:
      summary: Add an entry to the leaderboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - score
              properties:
                username:
                  type: string
                  example: "Vansh"
                score:
                  type: integer
                  example: 280
      responses:
        '201':
          description: Entry added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      score:
                        type: integer
        '400':
          description: Missing or invalid fields

  /remove:
    delete:
      summary: Remove an entry from the leaderboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                username:
                  type: string
                  example: "Vansh"
      responses:
        '200':
          description: Entry removed successfully
        '404':
          description: Entry not found

  /leaderboard:
    get:
      summary: Returns the top 10 leaderboard positions
      responses:
        '200':
          description: Top 10 scores
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        username:
                          type: string
                        score:
                          type: integer
                        created_at:
                          type: string
                          format: date-time

  /info:
    get:
      summary: Returns quantitative statistics for all entries
      responses:
        '200':
          description: Statistics including mean, median, quartiles, std deviation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      count:
                        type: integer
                      mean:
                        type: number
                      median:
                        type: number
                      min:
                        type: integer
                      max:
                        type: integer
                      q1:
                        type: number
                      q3:
                        type: number
                      std_deviation:
                        type: number
                      range:
                        type: integer

  /performance:
    get:
      summary: Returns average endpoint execution time in milliseconds
      responses:
        '200':
          description: Performance metrics per endpoint
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    additionalProperties:
                      type: object
                      properties:
                        avg_ms:
                          type: number
                        calls:
                          type: integer
                        min_ms:
                          type: integer
                        max_ms:
                          type: integer

  /history:
    get:
      summary: Returns all score submissions with timestamps (bonus endpoint)
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
          example: "2025-01-01"
        - name: to
          in: query
          schema:
            type: string
            format: date
          example: "2025-12-31"
        - name: username
          in: query
          schema:
            type: string
          example: "Vansh"
      responses:
        '200':
          description: Filtered history of all submissions